сделал крч все совсем правильно, добавил постконструктор
