+++ src/main/java/pro/sky/block3/controllers/IngridientsController.java
@@ -9,6 +9,7 @@ import pro.sky.block3.services.IngridientServices;
 import pro.sky.block3.services.RecieptsServices;
 
 
+import javax.annotation.PostConstruct;
 import java.util.ArrayList;
 
 @RestController
@@ -22,6 +23,13 @@ public class IngridientsController {
         this.ingridientServices = ingridientServices;
     }
 
+    @PostConstruct
+            void first() {
+        ingridientServices.createIngridient("яйца", 2, "шт.");
+        ingridientServices.createIngridient("молоко", 0.5, "л.");
+        ingridientServices.createIngridient("курица", 200, "гр.");
+        ingridientServices.createIngridient("соль", 1, "по вкусу");
+    }
     @GetMapping("/create")
     public String createIng(String name, double count, String unit) {
         return ingridientServices.createInController(name, count, unit);
